Add asset rename utilities and tests
